test(push): move the opt-out repro into the Xcode test target so CI runs it (#746)
